test: move perMessageDeflate tests to regression/issue/29684
The pre-existing WebSocket tests in test/js/first_party/ws/ws.test.ts
have ~30 1000ms-timeout flakes under ASAN/debug that obscure whether
this PR's tests actually exercise the fix. Move the new cases to a
dedicated regression file where the gate can see a clean without-fix
(8 fail) vs with-fix (0 fail) signal.
